Arregla tiempo de medir de RTAStar #5

+++ practica_1/src/src/tracks/singlePlayer/evaluacion/src_CANO_CAMARERO_BLANCA/AgenteRTAStar.java
@@ -24,7 +24,7 @@ public class AgenteRTAStar extends AbstractPlayer {
 
     // Variables para Act y métricas
     Boolean planCalculado = false;
-    int runtime = 0;
+    double runtime = 0;
     int tam_plan = 0;
 
     // Variables auxiliares relacionadas con coordenadas
